refactor(viewer): modify Viewer style location
