+++ libs/insight-viewer/src/Annotation/components/TextViewer/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { textStyle, svgBoxStyle, TEXT_SIZE, LINE_HEIGHT } from '../../../Viewer/Viewer.styles'
+import { textStyle, svgBoxStyle, TEXT_SIZE, LINE_HEIGHT } from '../../viewer.styles'
 import { TEXT_PADDING } from '../../../const'
 
 import type { TextViewerProps } from './TextViewer.types'
