+++ libs/insight-viewer/src/Annotation/viewer.styles.ts
